pyproject.toml
[project]
name = "mecagent-technical-test"
version = "0.1.0"
description = "Just a technical test"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cadquery>=2.5.2",
    "datasets>=3.6.0",
    "ipykernel>=6.29.5",
    "scipy>=1.15.3",
    "trimesh>=4.6.11",
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "accelerate>=0.25.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "tqdm>=4.65.0",
    "numpy>=1.24.0",
]
